requirements.txt
asgiref==3.9.2
attrs==25.3.0
django==5.2.6
django-cors-headers==4.9.0
django-filter==23.3
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
pyyaml==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.2.0
validators==0.35.0
